Add a way to generate a MANIFEST file for directories that don't have one.
HOWEVER: this got the testing phase far enough along to hit
net::ERR_INSUFFICIENT_RESOURCES
syncing a feed directory that currently occupies 67M.
which, I think, tells me that a progressive web app isn't going
to work for this, at least not if I want to be able to store
books and such.

Sigh.